[vbox] Simplify vbox-export-snapshots
This commit simplifies `vbox-export-snapshots.py` to support exporting
only a single snapshot at a time. This removes the complexity of
managing multiple snapshots and configuration files.

The original plan was to use `vbox-export-snapshots.py` to export
several snapshot with FLARE-VM installed using `vbox-build-vm.py`.
However, this proved problematic because `vbox-build-vm.py` requires
executing `VM-Set-Legal-Notice`, which prevents the VM from starting
without manual interaction (to accept the legal notice).  Starting the
VM is necessary after setting the network interface to `hostonly` for
the internet detector to update the VM status.

To address this, the intended future approach is to handle both setting
the interface to `hostonly` and exporting the snapshot directly within
`vbox-build-vm.py`.  Because `vbox-export-snapshots.py` is now
independent of `vbox-build-vm.py`, it can be significantly simplified as
done in this commit.  Both scripts will share code that will be moved to
`vboxcommon.py` to avoid code duplication.
