--- virtualbox/configs/export_remnux.json
@@ -1,11 +1,0 @@
-{
-  "VM_NAME": "REMnux.testing",
-  "EXPORTED_VM_NAME": "REMnux",
-  "SNAPSHOTS": [
-    [
-      "DONE",
-      ".dynamic",
-      "REMnux (based on Ubuntu) with improved configuration"
-    ]
-  ]
-}
