--- virtualbox/configs/export_win10_flare-vm.json
@@ -1,21 +1,0 @@
-{
-  "VM_NAME": "FLARE-VM.testing",
-  "EXPORTED_VM_NAME": "FLARE-VM",
-  "SNAPSHOTS": [
-    [
-      "FLARE-VM",
-      ".dynamic",
-      "Windows 10 VM with FLARE-VM default configuration"
-    ],
-    [
-      "FLARE-VM.full",
-      ".full.dynamic",
-      "Windows 10 VM with FLARE-VM default configuration + visualstudio.vm + pdbs.pdbresym.vm + microsoft-office.vm"
-    ],
-    [
-      "FLARE-VM.EDU",
-      ".EDU",
-      "Windows 10 VM with FLARE-VM default configuration + FLARE-EDU materials"
-    ]
-  ]
-}
